+++ src/Repository/CaseEventRepository.php
@@ -4,10 +4,8 @@ namespace App\Repository;
 
 use App\Entity\CaseEntity;
 use App\Entity\CaseEvent;
-use App\Entity\CaseEventPartyRelation;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Common\Collections\Criteria;
-use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
 
@@ -33,14 +31,9 @@ class CaseEventRepository extends ServiceEntityRepository
 
     public function createAvailableCaseEventsForCaseQueryBuilder($alias, CaseEntity $caseEntity): QueryBuilder
     {
-        $relationAlias = $alias.'_relation';
-        $partyAlias = $relationAlias.'_part';
-
         return $this->createQueryBuilder($alias)
             ->where($alias.'.caseEntity = :case')
             ->setParameter('case', $caseEntity->getId(), 'uuid')
-            ->leftJoin(CaseEventPartyRelation::class, $relationAlias, Join::WITH, $alias.'.id = '.$relationAlias.'.caseEvent')
-            ->leftJoin($relationAlias.'.party', $partyAlias)
             ->orderBy($alias.'.receivedAt', Criteria::DESC)
             ;
     }