+++ CHANGELOG.md
@@ -15,6 +15,9 @@ about writing changes to this log.
   Marked `Document.originalFileName` as nullable
 - [TVIST1-743](https://jira.itkdev.dk/browse/TVIST1-743)
   Added case events.
+- [TVIST1-788](https://jira.itkdev.dk/browse/TVIST1-788)
+  - Updated `CaseEvent` with `senders` and `recipients` property.
+  - Removed `CaseEventPartyRelation`.
 
 ## [1.1.2] 2023-02-17